+++ main_project_code/TeamProject/iCollections/wwwroot/css/site.css
@@ -219,7 +219,7 @@ h1 {
 }
 
 .big-name {
-  font-size: 20px;
+  font-size: 15px;
   font-weight: bold;
 }
